requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
django-cors-headers
Pillow
